+++ src/main/java/com/flemmli97/improvedmobs/ai/ILadderFlagNode.java
@@ -1,8 +1,0 @@
-package com.flemmli97.improvedmobs.ai;
-
-public interface ILadderFlagNode {
-
-    void setCanClimbLadder(boolean flag);
-
-    boolean canClimbLadder();
-}