+++ src/main/java/com/flemmli97/improvedmobs/utils/INodeBreakable.java
@@ -1,0 +1,12 @@
+package com.flemmli97.improvedmobs.utils;
+
+public interface INodeBreakable {
+
+    void setCanBreakBlocks(boolean flag);
+
+    boolean canBreakBlocks();
+
+    void setCanClimbLadder(boolean flag);
+
+    boolean canClimbLadder();
+}